+++ src/components/layouts/Header/index.jsx
@@ -1,8 +1,12 @@
 import { Link } from 'react-router-dom';
+import SearchMenu from '@/components/pages/home/SearchMenu';
 
 import { HeaderBox, HomeButton } from './styles';
+import useLogics from './logics';
 
 export default function Header() {
+  const { menu, handleChangeMenu, handleClickSearch } = useLogics();
+
   return (
     <HeaderBox>
       <Link to="/">
@@ -10,6 +14,11 @@ export default function Header() {
           TMI
         </HomeButton>
       </Link>
+      <SearchMenu
+        menu={menu}
+        onChange={handleChangeMenu}
+        onClick={handleClickSearch}
+      />
     </HeaderBox>
   );
 }
