+++ src/components/pages/home/SearchMenu/index.jsx
@@ -1,5 +1,8 @@
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { solid } from '@fortawesome/fontawesome-svg-core/import.macro';
+
 import {
-  SearchMenuBox, SubTitle, MenuInputBox, MenuInput, SubmitButton,
+  SearchMenuBox, MenuInputBox, MenuInput, SubmitButton,
 } from './styles';
 
 export default function SearchMenu({
@@ -9,9 +12,6 @@ export default function SearchMenu({
 }) {
   return (
     <SearchMenuBox>
-      <SubTitle>
-        생각나는 메뉴를 알려주세요!
-      </SubTitle>
       <MenuInputBox>
         <MenuInput
           type="text"
@@ -19,8 +19,10 @@ export default function SearchMenu({
           onChange={onChange}
           placeholder="메뉴를 입력하세요"
         />
-        <SubmitButton type="button" onClick={onClick}>
-          검색
+        <SubmitButton role="search" type="button" onClick={onClick}>
+          <FontAwesomeIcon icon={solid('magnifying-glass')}>
+            검색
+          </FontAwesomeIcon>
         </SubmitButton>
       </MenuInputBox>
     </SearchMenuBox>
