+++ src/components/pages/home/SearchMenu/index.test.jsx
@@ -10,13 +10,6 @@ describe('SearchMenu', () => {
     <SearchMenu onChange={handleChange} onClick={handleClick} />,
   );
 
-  it('서브 타이틀이 화면에 나타난다.', () => {
-    const subTitle = '생각나는 메뉴를 알려주세요!';
-    const { getByText } = renderSearchMenu();
-
-    expect(getByText(subTitle)).not.toBeNull();
-  });
-
   it('onChange 이벤트 핸들러가 실행된다.', () => {
     const placeholderText = '메뉴를 입력하세요';
     const { getByPlaceholderText } = renderSearchMenu();
@@ -27,9 +20,9 @@ describe('SearchMenu', () => {
   });
 
   it('onClick 이벤트 핸들러가 실행된다.', () => {
-    const { getByText } = renderSearchMenu();
+    const { getByRole } = renderSearchMenu();
 
-    fireEvent.click(getByText('검색'));
+    fireEvent.click(getByRole('search'));
 
     expect(handleClick).toBeCalled();
   });
